# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Indexer APIs for Andamio dapp-system.",
        "title": "Andamio Indexer 1.0.0",
        "contact": {
            "name": "Andamio Support",
            "url": "https://www.andamio.com/support",
            "email": "dev@andamio.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "0.0.0.0:42069",
    "basePath": "/api/v1/indexer",
    "paths": {
...
        "/assets/{asset_fingerprint}/utxos": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of UTxOs containing a specific asset fingerprint, with support for pagination.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Get UTxOs by Asset Fingerprint",
                "operationId": "getUTxOsByAssetFingerprint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The asset fingerprint (hex-encoded) to retrieve UTxOs for.",
                        "name": "asset_fingerprint",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Maximum number of results to return.",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of results to skip.",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved UTxOs.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/viewmodel.SimpleUTxO"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid asset fingerprint or pagination parameters.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asset fingerprint not found or no UTxOs found.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/datums/{datum_hash}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve details of a specific datum by its hash.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Datums"
                ],
                "summary": "Get Datum by Hash",
                "operationId": "getDatumByHash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The datum hash (hex-encoded) to retrieve.",
                        "name": "datum_hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved datum.",
                        "schema": {
                            "$ref": "#/definitions/viewmodel.Datum"
                        }
                    },
                    "400": {
                        "description": "Invalid datum hash.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Datum not found.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/datums/{datum_hash}/transactions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves a list of transactions that reference a specific datum hash.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Datums"
                ],
                "summary": "Get Transactions by Datum Hash",
                "operationId": "getTransactionsByDatumHash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The datum hash (hex-encoded) to retrieve transactions for.",
                        "name": "datum_hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved transactions.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/viewmodel.Transaction"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid datum hash.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Datum hash not found or no transactions found.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/indexer/assets/token/{tokenname}/transactions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves transactions associated with a given token name with pagination.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Get Transactions by Token Name",
                "operationId": "getTransactionsByTokenName",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The token name to retrieve transactions for (hex-encoded).",
                        "name": "tokenname",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Maximum number of results to return.",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of results to skip.",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved transactions.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/viewmodel.Transaction"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid token name or pagination parameters.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Token name not found or no transactions found.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "viewmodel.Asset": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "fingerprint": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "name_hex": {
                    "type": "string"
                },
                "policy_id": {
                    "type": "string"
                },
                "utxo_id": {
                    "type": "string"
                },
                "utxo_index": {
                    "type": "integer"
                }
            }
        },
        "viewmodel.Datum": {
            "type": "object",
            "properties": {
                "datum_cbor": {
                    "type": "string"
                },
                "datum_hash": {
                    "type": "string"
                },
                "utxo_id": {
                    "type": "string"
                },
                "utxo_index": {
                    "type": "integer"
                }
            }
        },
        "viewmodel.Redeemer": {
            "type": "object",
            "properties": {
                "cbor": {
                    "description": "CBOR string representation",
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "tag": {
                    "type": "integer"
                },
                "transaction_hash": {
                    "type": "string"
                }
            }
        },
        "viewmodel.SimpleUTxO": {
            "type": "object",
            "properties": {
                "transaction_hash": {
                    "type": "string"
                },
                "utxo_id": {
                    "type": "string"
                },
                "utxo_index": {
                    "type": "integer"
                }
            }
        },
        "viewmodel.Transaction": {
            "type": "object",
            "properties": {
                "block_hash": {
                    "type": "string"
                },
                "block_number": {
                    "type": "integer"
                },
                "certificates": {
                    "description": "Slice of CBOR string representations",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fee": {
                    "type": "integer"
                },
                "inputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/viewmodel.TransactionInput"
                    }
                },
                "metadata": {
                    "description": "CBOR string representation",
                    "type": "string"
                },
                "outputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/viewmodel.TransactionOutput"
                    }
                },
                "reference_inputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/viewmodel.SimpleUTxO"
                    }
                },
                "slot_number": {
                    "type": "integer"
                },
                "transaction_hash": {
                    "type": "string"
                },
                "ttl": {
                    "type": "integer"
                },
                "withdrawals": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "witness": {
                    "$ref": "#/definitions/viewmodel.Witness"
                }
            }
        },
        "viewmodel.TransactionInput": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "amount": {
                    "type": "integer"
                },
                "asset": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/viewmodel.Asset"
                    }
                },
                "cbor": {
                    "description": "CBOR string representation",
                    "type": "string"
                },
                "datum": {
                    "$ref": "#/definitions/viewmodel.Datum"
                },
                "transaction_hash": {
                    "type": "string"
                },
                "utxo_id": {
                    "type": "string"
                },
                "utxo_index": {
                    "type": "integer"
                }
            }
        },
        "viewmodel.TransactionOutput": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "amount": {
                    "type": "integer"
                },
                "asset": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/viewmodel.Asset"
                    }
                },
                "cbor": {
                    "description": "CBOR string representation",
                    "type": "string"
                },
                "datum": {
                    "$ref": "#/definitions/viewmodel.Datum"
                },
                "utxo_id": {
                    "type": "string"
                },
                "utxo_index": {
                    "type": "integer"
                }
            }
        },
...
        "viewmodel.Witness": {
            "type": "object",
            "properties": {
                "plutus_data": {
                    "description": "Slice of CBOR string representations",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "plutus_v1_scripts": {
                    "description": "Slice of CBOR string representations",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "plutus_v2_scripts": {
                    "description": "Slice of CBOR string representations",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "plutus_v3_scripts": {
                    "description": "Slice of CBOR string representations",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "redeemers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/viewmodel.Redeemer"
                    }
                },
                "transaction_hash": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}